backend/generate-textures/requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
requests==2.31.0
